+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Features
+
+- Findex Cloud beta
+
+---
+
 ## [4.0.1] - 2023-02-17
 
 ### Bug Fixes